refactor: more cleanup
