+++ src/types/mod.rs
@@ -3,11 +3,13 @@ mod enc_secret_params;
 mod key;
 mod key_id;
 mod packet;
+mod plain_secret_params;
 mod public_key;
 mod revocation_key;
 mod s2k;
 mod secret_key;
 mod secret_key_repr;
+mod secret_params;
 mod user;
 
 pub use self::compression::*;
@@ -15,9 +17,11 @@ pub use self::enc_secret_params::*;
 pub use self::key::*;
 pub use self::key_id::*;
 pub use self::packet::*;
+pub use self::plain_secret_params::*;
 pub use self::public_key::*;
 pub use self::revocation_key::*;
 pub use self::s2k::*;
 pub use self::secret_key::*;
 pub use self::secret_key_repr::*;
+pub use self::secret_params::*;
 pub use self::user::*;
